fix date range picker and refactor
