+++ src/components/timebank/timebank-daterange-picker.tsx
@@ -1,17 +1,13 @@
 import { DatePicker } from "@mui/x-date-pickers";
-import { DateTime } from "luxon";
-import { useEffect, useState } from "react";
-import { DailyEntry } from "../../generated/client";
-import { Range } from "../../types";
+import { DateRange } from "../../types";
 import strings from "../../localization/strings";
 
 /**
  * Component properties
  */
 interface Props {
-  dailyEntries: DailyEntry[];
-  setSelectedEntries: (selectedEntries: DailyEntry[]) => void;
-  today: DateTime;
+  range: DateRange;
+  handleDateRangeChange: (range: DateRange) => void;
 }
 
 /**
@@ -19,40 +15,7 @@ interface Props {
  *
  * @param props Component properties
  */
-const DateRangePicker = ({ setSelectedEntries, dailyEntries, today }: Props) => {
-  const [range, setRange] = useState<Range>({
-    start: today.minus({ days: 7 }),
-    end: today
-  });
-
-  useEffect(() => {
-    getDateRangeEntries();
-  }, [range]);
-
-  /**
-   * Gets daily entries within time range
-   * Filters null entries, commonly weekends.
-   */
-  const getDateRangeEntries = () => {
-    if (range.start && range.end) {
-      const selectedDays = range.end.diff(range.start, "days").toObject();
-      const result = [];
-
-      for (let i = 0; selectedDays.days && i <= selectedDays.days; i++) {
-        result.push(
-          dailyEntries.filter(
-            (item) =>
-              item.logged &&
-              item.expected &&
-              DateTime.fromJSDate(item.date).toISODate() ===
-                range.start?.plus({ days: i }).toISODate()
-          )[0]
-        );
-      }
-      setSelectedEntries(result.filter((item) => item));
-    }
-  };
-
+const DateRangePicker = ({ handleDateRangeChange, range }: Props) => {
   return (
     <>
       <DatePicker
@@ -61,9 +24,9 @@ const DateRangePicker = ({ setSelectedEntries, dailyEntries, today }: Props) => 
           mx: "1%"
         }}
         label={strings.timeExpressions.startDate}
-        onChange={(dateTime) => setRange({ ...range, start: dateTime })}
+        onChange={(dateTime) => dateTime && handleDateRangeChange({ ...range, start: dateTime })}
         value={range.start}
-        maxDate={range.end?.minus({ days: 1 })}
+        maxDate={range.end}
       />
       <DatePicker
         sx={{
@@ -71,9 +34,9 @@ const DateRangePicker = ({ setSelectedEntries, dailyEntries, today }: Props) => 
           mx: "1%"
         }}
         label={strings.timeExpressions.endDate}
-        onChange={(dateTime) => setRange({ ...range, end: dateTime })}
+        onChange={(dateTime) => dateTime && handleDateRangeChange({ ...range, end: dateTime })}
         value={range.end}
-        minDate={range.start?.plus({ days: 1 })}
+        minDate={range.start}
       />
     </>
   );
